# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/annotations": {
            "get": {
                "description": "list annotation entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List annotation entities",
                "operationId": "list-annotation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Annotation"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create annotation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create annotation",
                "operationId": "create-annotation",
                "parameters": [
                    {
                        "description": "Annotation entity",
                        "name": "annotation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Annotation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Annotation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/annotations/{id}": {
            "get": {
                "description": "get annotation by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a annotation entity by ID",
                "operationId": "get-annotation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Annotation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Annotation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update annotation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a annotation entity by ID",
                "operationId": "update-annotation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Annotation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Annotation entity",
                        "name": "annotation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Annotation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Annotation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get annotation by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a annotation entity by ID",
                "operationId": "delete-annotation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Annotation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Annotation": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AnnotationQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.AnnotationEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "messages": {
                    "description": "Messages holds the value of the \"messages\" field.",
                    "type": "string"
                }
            }
        },
        "ent.AnnotationEdges": {
            "type": "object",
            "properties": {
                "dispensemedicine": {
                    "description": "Dispensemedicine holds the value of the dispensemedicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DispenseMedicine"
                    }
                }
            }
        },
        "ent.Bill": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "annotation": {
                    "description": "Annotation holds the value of the \"annotation\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BillQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BillEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BillEdges": {
            "type": "object",
            "properties": {
                "dispenseMedicines": {
                    "description": "DispenseMedicines holds the value of the dispenseMedicines edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DispenseMedicine"
                },
                "payments": {
                    "description": "Payments holds the value of the payments edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Payment"
                },
                "pharmacists": {
                    "description": "Pharmacists holds the value of the pharmacists edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pharmacist"
                }
            }
        },
...
        "ent.DispenseMedicine": {
            "type": "object",
            "properties": {
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DispenseMedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DispenseMedicineEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DispenseMedicineEdges": {
            "type": "object",
            "properties": {
                "annotation": {
                    "description": "Annotation holds the value of the annotation edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Annotation"
                },
                "bills": {
                    "description": "Bills holds the value of the Bills edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bill"
                },
                "pharmacist": {
                    "description": "Pharmacist holds the value of the pharmacist edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pharmacist"
                },
                "prescription": {
                    "description": "Prescription holds the value of the prescription edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prescription"
                }
            }
        },
        "ent.Doctor": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DoctorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DoctorEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.DoctorEdges": {
            "type": "object",
            "properties": {
                "doctorprescription": {
                    "description": "Doctorprescription holds the value of the doctorprescription edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                }
            }
        },
        "ent.DrugAllergy": {
            "type": "object",
            "properties": {
                "dateTime": {
                    "description": "DateTime holds the value of the \"dateTime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DrugAllergyQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DrugAllergyEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DrugAllergyEdges": {
            "type": "object",
            "properties": {
                "medicine": {
                    "description": "Medicine holds the value of the medicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicine"
                },
                "patient": {
                    "description": "Patient holds the value of the patient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientInfo"
                },
                "pharmacist": {
                    "description": "Pharmacist holds the value of the pharmacist edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pharmacist"
                }
            }
        },
...
        "ent.Medicine": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "brand": {
                    "description": "Brand holds the value of the \"brand\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the MedicineQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineEdges"
                },
                "howtouse": {
                    "description": "Howtouse holds the value of the \"howtouse\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "price": {
                    "description": "Price holds the value of the \"price\" field.",
                    "type": "number"
                },
                "serial": {
                    "description": "Serial holds the value of the \"serial\" field.",
                    "type": "string"
                }
            }
        },
        "ent.MedicineEdges": {
            "type": "object",
            "properties": {
                "drugallergys": {
                    "description": "Drugallergys holds the value of the drugallergys edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DrugAllergy"
                    }
                },
                "levelOfDangerous": {
                    "description": "LevelOfDangerous holds the value of the LevelOfDangerous edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.LevelOfDangerous"
                },
                "medicineType": {
                    "description": "MedicineType holds the value of the MedicineType edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.MedicineType"
                },
                "medicinepresciption": {
                    "description": "Medicinepresciption holds the value of the medicinepresciption edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Prescription"
                    }
                },
                "order": {
                    "description": "Order holds the value of the order edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Order"
                    }
                },
                "unitOfMedicine": {
                    "description": "UnitOfMedicine holds the value of the UnitOfMedicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UnitOfMedicine"
                }
            }
        },
...
        "ent.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "integer"
                },
                "datetime": {
                    "description": "Datetime holds the value of the \"datetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OrderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OrderEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "price": {
                    "description": "Price holds the value of the \"price\" field.",
                    "type": "number"
                },
                "total": {
                    "description": "Total holds the value of the \"total\" field.",
                    "type": "number"
                }
            }
        },
        "ent.OrderEdges": {
            "type": "object",
            "properties": {
                "company": {
                    "description": "Company holds the value of the company edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Company"
                    }
                },
                "medicine": {
                    "description": "Medicine holds the value of the medicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Medicine"
                    }
                },
                "pharmacist": {
                    "description": "Pharmacist holds the value of the pharmacist edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Pharmacist"
                }
            }
        },
        "ent.PatientInfo": {
            "type": "object",
            "properties": {
                "age": {
                    "description": "Age holds the value of the \"age\" field.",
                    "type": "integer"
                },
                "cardNumber": {
                    "description": "CardNumber holds the value of the \"cardNumber\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PatientInfoQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientInfoEdges"
                },
                "gender": {
                    "description": "Gender holds the value of the \"gender\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Payment": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaymentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaymentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "pay": {
                    "description": "Pay holds the value of the \"pay\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PaymentEdges": {
            "type": "object",
            "properties": {
                "bills": {
                    "description": "Bills holds the value of the Bills edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bill"
                    }
                }
            }
        },
        "ent.Pharmacist": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PharmacistQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PharmacistEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "password": {
                    "description": "Password holds the value of the \"password\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PharmacistEdges": {
            "type": "object",
            "properties": {
                "bills": {
                    "description": "Bills holds the value of the Bills edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bill"
                    }
                },
                "dispensemedicine": {
                    "description": "Dispensemedicine holds the value of the dispensemedicine edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DispenseMedicine"
                    }
                },
                "drugallergys": {
                    "description": "Drugallergys holds the value of the drugallergys edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.DrugAllergy"
                    }
                },
                "order": {
                    "description": "Order holds the value of the order edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Order"
                    }
                }
            }
        },
        "ent.Prescription": {
            "type": "object",
            "properties": {
                "Value": {
                    "description": "Value holds the value of the \"Value\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrescriptionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrescriptionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PrescriptionEdges": {
            "type": "object",
            "properties": {
                "dispensemedicine": {
                    "description": "Dispensemedicine holds the value of the dispensemedicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DispenseMedicine"
                },
                "prescriptiondoctor": {
                    "description": "Prescriptiondoctor holds the value of the prescriptiondoctor edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Doctor"
                },
                "prescriptionmedicine": {
                    "description": "Prescriptionmedicine holds the value of the prescriptionmedicine edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Medicine"
                },
                "prescriptionpatient": {
                    "description": "Prescriptionpatient holds the value of the prescriptionpatient edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PatientInfo"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}